完成删除文章功能并重构代码

+++ src/main/java/com/modespring/core/service/impl/NodeServiceImpl.java
@@ -6,6 +6,7 @@ import com.modespring.core.service.NodeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -23,8 +24,28 @@ public class NodeServiceImpl implements NodeService {
         return nodeDao.save(node);
     }
 
-    public void delete(Integer id) {
-        nodeDao.delete(id);
+    public Boolean hasArticle(Integer id) {
+        List<Node> nodeList = nodeDao.findAll();
+        return nodeList.size() > 0;
+    }
+
+    public void delete(Integer id) throws Exception {
+        if (this.hasArticle(id)) {
+            throw new Exception("该栏目下有未删除的文章");
+        } else {
+            nodeDao.delete(id);
+        }
+    }
+
+    @Override
+    public void deleteAll(Integer[] id) throws Exception {
+        for (int i = 0; id != null && i < id.length; i++) {
+            if (this.hasArticle(id[i])) {
+                throw new Exception("该栏目下有未删除的文章");
+            } else {
+                nodeDao.delete(id[i]);
+            }
+        }
     }
 
     public Node update(Node node) {
@@ -32,6 +53,18 @@ public class NodeServiceImpl implements NodeService {
     }
 
     public List<Node> updateALL(List<Node> nodeList) {
+        return nodeDao.save(nodeList);
+    }
+
+    @Override
+    public List<Node> updateALL(Integer[] id, String[] name, String[] title) {
+        List<Node> nodeList = new ArrayList<Node>();
+        for (int i = 0; i < id.length; i++) {
+            Node node = nodeDao.getOne(id[i]);
+            node.setName(name[i]);
+            node.setTitle(title[i]);
+            nodeList.add(node);
+        }
         return nodeDao.save(nodeList);
     }
